Reminder CRUD

+++ app/src/main/java/sg/edu/nus/iss/se/ft05/medipal/Category.java
@@ -65,7 +65,7 @@ public class Category {
         return categoryDAO.insert(this);
     }
 
-    public long update(Context context){
+    public int update(Context context){
         categoryDAO = new CategoryDAOImpl(context);
         return categoryDAO.update(this);
     }
