+++ app/src/main/java/sg/edu/nus/iss/se/ft05/medipal/Medicine.java
@@ -157,7 +157,7 @@ public class Medicine {
         return medicineDAO.insert(this);
     }
 
-    public long update(Context context){
+    public int update(Context context){
         medicineDAO = new MedicineDAOImpl(context);
         return medicineDAO.update(this);
     }
